+++ src/Summoner/Text.hs
@@ -2,6 +2,7 @@ module Summoner.Text
        ( packageToModule
        , intercalateMap
        , headToUpper
+       , tconcatMap
        ) where
 
 import qualified Data.Char as C
@@ -10,7 +11,7 @@ import qualified Data.Text as T
 -- | Creates module name from the name of the package
 -- Ex: @my-lovely-project@ — @MyLovelyProject@
 packageToModule :: Text -> Text
-packageToModule = T.concat . map headToUpper . T.splitOn "-"
+packageToModule = tconcatMap headToUpper . T.splitOn "-"
 
 -- | Converts every element of list into 'Text' and then joins every element
 -- into single 'Text' like 'T.intercalate'.
@@ -21,3 +22,7 @@ headToUpper :: Text -> Text
 headToUpper t = case T.uncons t of
     Nothing      -> ""
     Just (x, xs) -> T.cons (C.toUpper x) xs
+
+-- | Convert every element of a list into text, and squash the results
+tconcatMap :: (a -> Text) -> [a] -> Text
+tconcatMap f = T.concat . map f